python_spring_work_2022/unit_three/TestBot/requirements.txt
aiofile==3.7.4
aiohttp==3.8.1
aiosignal==1.2.0
art==5.6
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
caio==0.9.5
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
fonttools==4.33.3
forex-python==1.8
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
keyboard==0.13.5
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pdfrw==0.4
Pillow==9.1.0
pluggy==1.0.0
psycopg==3.0.14
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
pytest==7.1.2
python-box==6.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
self==2020.12.3
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.9
voluptuous==0.13.1
yandex-weather-api==0.2.4
yarl==1.7.2
